+++ src/port_diff/extract.rs
@@ -1,5 +1,4 @@
-use crate::PortDiffGraph;
-use itertools::Itertools;
+use crate::{graph_view::MergeStrategy, PortDiffGraph};
 
 use crate::{Graph, PortDiff};
 
@@ -11,44 +10,29 @@ impl<G: Graph> PortDiff<G> {
     where
         G: 'a,
     {
-        let graphs = diffs
+        Self::try_merge(diffs.into_iter().map(|d| d.clone())).is_ok()
+    }
+
+    pub fn try_merge(
+        diffs: impl IntoIterator<Item = PortDiff<G>>,
+    ) -> Result<PortDiffGraph<G>, IncompatiblePortDiff> {
+        let mut graphs = diffs
             .into_iter()
-            .map(|d| PortDiffGraph::from_sinks(vec![d.clone()]))
-            .collect_vec();
+            .map(|d| PortDiffGraph::from_sinks(vec![d]));
 
-        if graphs.is_empty() {
-            return true;
-        }
+        // cannot use reduce here as we need to handle error
+        let Some(fst) = graphs.next() else {
+            return Ok(PortDiffGraph::default());
+        };
 
-        let merged_graph = graphs
-            .clone()
-            .into_iter()
-            .reduce(|mut g1, g2| {
-                g1.merge(g2);
-                g1
-            })
-            .expect("not empty");
-
-        // For each node that is an ancestor of two or more graphs...
-        // TODO: this does not work
-        // for diff_ptr in GraphView::lowest_common_ancestors(&graphs) {
-        // Check that its outgoing edges are compatible
-        // (this must hold everywhere, but already holds elsewhere as the
-        // set of outgoing edges in non-lca nodes remains unchanged).
-        //     let edges = merged_graph.inner().edges(diff_ptr.into()).collect_vec();
-        //     if !EdgeData::are_compatible(edges.iter().map(|e| e.weight())) {
-        //         return false;
-        //     }
-        // }
-        merged_graph.is_squashable()
+        graphs.try_fold(fst, |mut g1, g2| {
+            g1.merge(g2, MergeStrategy::FailOnConflicts)?;
+            Ok(g1)
+        })
     }
 
     pub fn extract_graph(diffs: Vec<PortDiff<G>>) -> Result<G, IncompatiblePortDiff> {
-        if !PortDiff::are_compatible(&diffs) {
-            return Err(IncompatiblePortDiff);
-        }
-
-        let graph = PortDiffGraph::from_sinks(diffs);
+        let graph = Self::try_merge(diffs)?;
         let diff = PortDiff::squash(&graph);
         Ok(diff.try_unwrap_graph().unwrap())
     }
@@ -56,6 +40,7 @@ impl<G: Graph> PortDiff<G> {
 #[cfg(feature = "portgraph")]
 #[cfg(test)]
 mod tests {
+    use itertools::Itertools;
     use portgraph::render::DotFormat;
     use portgraph::PortView;
     use rstest::rstest;